+++ common/perlin/source/perlin.c
@@ -18,7 +18,7 @@
 #include <math.h>
 #include <stdlib.h>
 
-int generate_perlin_field(perlin_vec2_t * field, unsigned int width, unsigned int height) {
+int generate_perlin_field(perlin_vec2_t * field, size_t width, size_t height) {
     for(size_t y = 0; y < height; y++) {
         for(size_t x = 0; x < width; x++) {
             float angle = (float) random();
@@ -35,7 +35,7 @@ static int clip(int x, int limit) {
     return x - (limit * (x / limit));
 }
 
-static float dot_product_gradient(perlin_vec2_t * field, unsigned int x, unsigned int y, unsigned int width, unsigned int height, float dx, float dy) {
+static float dot_product_gradient(perlin_vec2_t * field, size_t x, size_t y, size_t width, size_t height, float dx, float dy) {
     perlin_vec2_t vec = field[(clip(y, height) * height) + clip(x, width)];
     return (vec.x * dx) + (vec.y * dy);
 }
@@ -46,16 +46,16 @@ static float interpolate(float i, float j, float w) {
     return ((j - i) * ((w * ((w * 6) - 15)) + 10) * w * w * w) + i;
 }
 
-float get_perlin_value(float x, float y, perlin_vec2_t * field, unsigned int width, unsigned int height) {
-    unsigned int ix = (unsigned int) floorf(x),
-                 iy = (unsigned int) floorf(y);
+float get_perlin_value(float x, float y, perlin_vec2_t * field, size_t width, size_t height) {
+    size_t ix = (size_t) floorf(x),
+           iy = (size_t) floorf(y);
     
     float xint[2];
     #pragma omp simd
-    for(unsigned int j = 0; j < 2; j++) {
+    for(size_t j = 0; j < 2; j++) {
         float dots[2];
         #pragma omp simd
-        for(unsigned int i = 0; i < 2; i++) {
+        for(size_t i = 0; i < 2; i++) {
             dots[i] = dot_product_gradient(field, ix+i, iy+j, width, height, x - ((float)(ix + i)), y - ((float)(iy + j)));
         }
         xint[j] = interpolate(dots[0], dots[1], x - ix);
@@ -65,7 +65,7 @@ float get_perlin_value(float x, float y, perlin_vec2_t * field, unsigned int wid
     return yint;
 }
 
-float get_perlin_value_fractal(float x, float y, float scale, unsigned int depth, perlin_vec2_t * field, unsigned int width, unsigned int height) {
+float get_perlin_value_fractal(float x, float y, float scale, unsigned int depth, perlin_vec2_t * field, size_t width, size_t height) {
     float frac = 0;
     for(unsigned int i = 0; i < depth; i++) {
         float iscale = scale * powf(2, i);
